refactor: extract session block and weekview component
